requirements.txt
pytorch-lightning==0.9.0
Pillow>=7.1.1
torch>=1.7.1
torchvision>=0.8.21
numpy
matplotlib
lmdb
tfrecord
imageio
h5py
wandb
